Added AddOtherAccountFlowTest. Refactored code.| #2038
